fixed from PVS-Studio:
V512 A call of the 'swprintf' function will lead to overflow of the buffer 'output'. colorpicker.cpp 545
V575 The number of processed elements should be passed to the 'swprintf' function as the second argument instead of buffer's size in bytes. colorpicker.cpp 545
V547 Expression 't0 > 0xFF' is always false. colorpicker.cpp 565
V557 Array overrun is possible. The value of 'i + 1' index could reach 50. nppqcp.cpp 597, 945
V796 It is possible that 'break' statement is missing in switch statement. nppqcp.cpp 354